requirements.txt
streamlit~=1.37.1
pillow~=10.4.0
python-dotenv~=0.21.0
appdirs~=1.4.4
fonttools~=4.51.0
markdown~=3.4.1
dask-expr~=1.1.13
pytesseract~=0.3.13